Hooking up to Test. Kinda.
